develop -> main

+++ src/main/java/com/devvictor/ecommerce_api/application/dtos/input/UpdateProductInputDTO.java
@@ -2,10 +2,10 @@ package com.devvictor.ecommerce_api.application.dtos.input;
 
 public record UpdateProductInputDTO(
         String productId,
-        Double price,
+        long price,
         String name,
         String description,
         String photoUrl,
-        Integer stockQuantity
+        int stockQuantity
 ) {
 }
